+++ packages/api-explorer/src/reducers/spec/utils.ts
@@ -126,7 +126,8 @@ export const fetchSpec = (key: string, specs: SpecList): SpecState => {
  * @param specs to use to find the default spec key
  */
 export const getSpecKey = (location: AbstractLocation, specs?: SpecList) => {
-  const pathNodes = location.pathname.split('/')
+  const pathName = location.pathname
+  const pathNodes = pathName.split('/')
   let specKey = ''
   if (
     pathNodes.length > 1 &&